backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
google-auth>=2.23.0
requests>=2.31.0
PyJWT>=2.8.0
